requirements.txt
fastapi
uvicorn[standard]
python-multipart
google-generativeai
python-dotenv
python-magic
faster-whisper
--extra-index-url https://download.pytorch.org/whl/cpu
torch